requirements.txt
Flask==3.1.2
requests==2.32.5
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.2.1
playwright==1.42.0
gunicorn==21.2.0
